requirements.txt
streamlit==1.25.0  # For building the web app interface
openai==0.27.8     # For interacting with the OpenAI API
python-dotenv==1.0.0  # For loading environment variables from a .env file